+++ lib/import_path.dart
@@ -1,0 +1,4 @@
+/// Import Path search library.
+library import_path;
+
+export 'src/import_path_base.dart';